refactor(slm-frontend): extract formatRelativeTime to shared dateUtils (#3314)
Creates src/utils/dateUtils.ts with a single exported formatRelativeTime
utility and removes the 5 private copies in FleetOverview, NodeLifecyclePanel,
NodeServicesPanel, ErrorMonitor, and SecurityView. No functional changes.
